Refactoring to only provide simple-tokenizer extensions which makes it easy to build syntax highlighting UI on top of
