+++ .eslintrc
@@ -2,6 +2,9 @@
     "parserOptions": {
         "ecmaVersion": 2015
     },
+    "env": {
+        "es6": true
+    },
     "globals": {
         define: true,
         require: true,
