+++ src/layer/text_util.js
@@ -3,4 +3,4 @@ const textTokens = new Set(["text", "rparen", "lparen"]);
 
 exports.isTextToken = function(tokenType) {
     return textTokens.has(tokenType);
-}
+};
