+++ src/tokenizer.js
@@ -1,6 +1,6 @@
 "use strict";
 const config = require("./config");
 
-const Tokenizer = require("./tokenizer_internal").Tokenizer;
-Tokenizer.prototype.reportError = config.reportError;
-exports.Tokenizer = Tokenizer;
+const TokenizerInternal = require("./tokenizer_internal").TokenizerInternal;
+TokenizerInternal.prototype.reportError = config.reportError;
+exports.Tokenizer = TokenizerInternal;
